Web Pad/service/website/requirements.txt
Flask>=2.2.2
pycryptodome>=3.15.0